pyproject.toml
[tool.poetry]
name = "project5000"
version = "0.1.0"
description = ""
authors = ["Евгений Ермак <144560375+EvgeniiErmak@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
django = "^5.0.2"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
